@@ -5,17 +5,20 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "@material-ui/core": "^4.0.0-rc.0",
-    "@material-ui/icons": "^4.2.1",
-    "prismjs": "1.15.0",
-    "react": "16.8.0",
-    "react-dom": "16.8.0",
-    "react-markdown": "^4.0.8",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "2.0.3",
-    "react-simple-code-editor": "0.9.1"
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^5.15.15",
+    "@mui/material": "^5.15.15",
+    "@mui/system": "^5.15.15",
+    "prismjs": "^1.29.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^8.0.7",
+    "react-router-dom": "^6.23.0",
+    "react-scripts": "5.0.1",
+    "react-simple-code-editor": "^0.13.1",
+    "yarn": "^1.22.22"
   },
-  "devDependencies": {},
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -27,5 +30,6 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "license": "MIT"
 }
