@@ -1,23 +1,25 @@
 import React from 'react';
-import Container from '@material-ui/core/Container';
-import TextField from '@material-ui/core/TextField';
-import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
-import Snackbar from '@material-ui/core/Snackbar';
-import { withRouter } from "react-router-dom";
+import Container from '@mui/material/Container';
+import TextField from '@mui/material/TextField';
+import Button from '@mui/material/Button';
+import Typography from '@mui/material/Typography';
+import Snackbar from '@mui/material/Snackbar';
+import { useNavigate } from "react-router-dom";
 
 class Register extends React.Component {
-    constructor(props){
+    constructor(props) {
         super(props);
         this.state = {
             username: '',
             password: '',
             error_message: '',
             show_error_message: false
-        }
+        };
     }
 
     handleSignUp(event) {
+        event.preventDefault();
+
         fetch('/api/auth/register', {
             method: 'POST',
             headers: {
@@ -25,8 +27,8 @@ class Register extends React.Component {
                 'Content-type': 'application/json'
             },
             body: JSON.stringify({
-                'username': this.state.username,
-                'password': this.state.password
+                username: this.state.username,
+                password: this.state.password
             })
         }).then((response) => {
             if (response.status === 400) {
@@ -34,7 +36,7 @@ class Register extends React.Component {
                     throw new Error(json.error);
                 });
             } else if (response.status === 200) {
-                this.props.history.push("/login");
+                this.props.navigate("/login");
             }
         }).catch((error) => {
             this.setState({
@@ -47,57 +49,57 @@ class Register extends React.Component {
     render() {
         return (
             <div>
-            <Container maxWidth="xs">
-                <Typography component="h1" variant="h2">
-                    Sign Up
-                </Typography>
-                <TextField
-                    variant="outlined"
-                    margin="normal"
-                    required
-                    fullWidth
-                    name="username"
-                    id="username"
-                    label="Username"
-                    autocomplete="username"
-                    autofocus
-                    onChange = {(event) => this.setState({username: event.target.value})}
+                <Container maxWidth="xs">
+                    <Typography component="h1" variant="h2">
+                        Sign Up
+                    </Typography>
+                    <TextField
+                        variant="outlined"
+                        margin="normal"
+                        required
+                        fullWidth
+                        name="username"
+                        id="username"
+                        label="Username"
+                        autoComplete="username"
+                        autoFocus
+                        onChange={(event) => this.setState({ username: event.target.value })}
+                    />
+                    <TextField
+                        variant="outlined"
+                        margin="normal"
+                        required
+                        fullWidth
+                        name="password"
+                        id="password"
+                        label="Password"
+                        type="password"
+                        onChange={(event) => this.setState({ password: event.target.value })}
+                    />
+                    <Button
+                        fullWidth
+                        variant="contained"
+                        color="primary"
+                        onClick={(event) => this.handleSignUp(event)}
+                    >
+                        Sign Up
+                    </Button>
+                </Container>
+                <Snackbar
+                    anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
+                    autoHideDuration={6000}
+                    message={this.state.error_message}
+                    open={this.state.show_error_message}
+                    onClose={() => this.setState({ show_error_message: false })}
                 />
-                <TextField
-                    variant="outlined"
-                    margin="normal"
-                    required
-                    fullWidth
-                    name="password"
-                    id="password"
-                    label="Password"
-                    type="password"
-                    onChange = {(event) => this.setState({password: event.target.value})}
-                />
-                <Button
-                    type="sumbit"
-                    fullWidth
-                    variant="contained"
-                    color="primary"
-                    onClick={(event) => this.handleSignUp(event)}
-                >
-                    Sign Up
-                </Button>
-            </Container>
-            <Snackbar
-                ref="snackbar"
-                anchorOrigin={{
-                    vertical: 'top',
-                    horizontal: 'center',
-                }}
-                autoHideDuration={6000}
-                message={this.state.error_message}
-                open={this.state.show_error_message}
-                onClose={(event, reason) => {this.setState({'show_error_message': false})}}
-            />
             </div>
         );
     }
 }
 
-export default withRouter(Register);
+function RegisterWithRouter(props) {
+    const navigate = useNavigate();
+    return <Register {...props} navigate={navigate} />;
+}
+
+export default RegisterWithRouter;
