@@ -1,12 +1,12 @@
 import React from 'react';
-import Container from '@material-ui/core/Container';
-import Typography from '@material-ui/core/Typography';
-import Table from '@material-ui/core/Table';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import TableCell from '@material-ui/core/TableCell';
-import TableBody from '@material-ui/core/TableBody';
-import DoneIcon from '@material-ui/icons/Done';
+import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
+import Table from '@mui/material/Table';
+import TableHead from '@mui/material/TableHead';
+import TableRow from '@mui/material/TableRow';
+import TableCell from '@mui/material/TableCell';
+import TableBody from '@mui/material/TableBody';
+import DoneIcon from '@mui/icons-material/Done';
 import { Link as RouterLink } from 'react-router-dom';
 
 const COMMON_TITLE = 'Statistics';
@@ -25,15 +25,15 @@ class Statistics extends React.Component {
         }
     }
 
-    fetchStatistics(courseSlug='') {
+    fetchStatistics(courseSlug = '') {
         if (courseSlug) {
-            fetch('/api/courses/'+courseSlug).then(function(response){
+            fetch('/api/courses/' + courseSlug).then(function (response) {
                 return response.json();
             }).then(
                 course => {
-                    this.setState({'title': COMMON_TITLE, 'abstract': course.title});
-            });
-            fetch('/api/statistics/courses/'+courseSlug).then(function(response){
+                    this.setState({ 'title': COMMON_TITLE, 'abstract': course.title });
+                });
+            fetch('/api/statistics/courses/' + courseSlug).then(function (response) {
                 return response.json();
             }).then(statistics => {
                 this.setState({
@@ -43,8 +43,8 @@ class Statistics extends React.Component {
                 });
             });
         } else {
-            this.setState({'title': COMMON_TITLE, 'abstract': COMMON_ABSTRACT});
-            fetch('/api/statistics').then(function(response){
+            this.setState({ 'title': COMMON_TITLE, 'abstract': COMMON_ABSTRACT });
+            fetch('/api/statistics').then(function (response) {
                 return response.json();
             }).then(statistics => {
                 this.setState({
@@ -56,13 +56,13 @@ class Statistics extends React.Component {
         }
     }
 
-    componentDidMount() {  
+    componentDidMount() {
         if (this.props.match.params.hasOwnProperty('courseSlug')) {
             this.fetchStatistics(this.props.match.params.courseSlug);
         } else {
             this.fetchStatistics();
         }
-        
+
     }
 
     componentWillReceiveProps(nextProps) {
@@ -93,13 +93,13 @@ class Statistics extends React.Component {
                                 <TableCell>Username</TableCell>
                                 {
                                     this.state.header.map(header => (
-                                    <TableCell align="right">
-                                        {
-                                            ( this.state.hasCourseSlug )
-                                            ? header['title']
-                                            : <RouterLink to={'/statistics/'+header['slug']}>{header['title']}</RouterLink> 
-                                        }
-                                    </TableCell>
+                                        <TableCell align="right">
+                                            {
+                                                (this.state.hasCourseSlug)
+                                                    ? header['title']
+                                                    : <RouterLink to={'/statistics/' + header['slug']}>{header['title']}</RouterLink>
+                                            }
+                                        </TableCell>
                                     ))
                                 }
                             </TableRow>
@@ -109,19 +109,19 @@ class Statistics extends React.Component {
                                 <TableRow key={row.uid}>
                                     <TableCell component="th" scope="row">{row.username}</TableCell>
                                     {
-                                        ( this.state.hasCourseSlug )
-                                        ? this.state.header.map(lesson => (
-                                            <TableCell align="right">
-                                                {
-                                                    (row.lessons[lesson['slug']].progress)
-                                                    && <DoneIcon/>
-                                                }
-                                            </TableCell>
+                                        (this.state.hasCourseSlug)
+                                            ? this.state.header.map(lesson => (
+                                                <TableCell align="right">
+                                                    {
+                                                        (row.lessons[lesson['slug']].progress)
+                                                        && <DoneIcon />
+                                                    }
+                                                </TableCell>
                                             ))
-                                        : this.state.header.map(course => (
-                                            <TableCell align="right">
-                                                {(row.courses[course['slug']].progress)}&nbsp;%
-                                            </TableCell>
+                                            : this.state.header.map(course => (
+                                                <TableCell align="right">
+                                                    {(row.courses[course['slug']].progress)}&nbsp;%
+                                                </TableCell>
                                             ))
                                     }
                                 </TableRow>
